Update documentation and fixed indexing oridunal_position
Documentation updated on feedback received. Now the dimension should be set to logical null + screenshot of how thats done in workbench
Table meta_template is now keyd on ordinal_position in addition to the existing databasename, tablename. This fixes the default listing so first column is now first on list etc.